% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{basic_mean_update}
\alias{basic_mean_update}
\title{basic_mean_update}
\usage{
basic_mean_update(prev_hidden_states, choices, variance_basic)
}
\arguments{
\item{prev_hidden_states}{a list structure containing the states from last round}

\item{choices}{a vector of [1] own choice and [2] opponent's choice from last round}

\item{variance_basic}{the updated estimate variance}
}
\value{
An updated mean estimate
}
\description{
0-ToM updates the mean of its parameter estimate
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
