% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{basic_p_op_1_fun}
\alias{basic_p_op_1_fun}
\title{basic_p_op_1_fun}
\usage{
basic_p_op_1_fun(hidden_states, params)
}
\arguments{
\item{hidden_states}{0-ToM's updated estimates for this round}
}
\value{
The estimated choice probability of the opponent
}
\description{
0-ToM combines the mean and variance of its parameter estimate into a final choice probability estimate.
NB: this is the function taken from the VBA package (Daunizeau 2014), which does not use 0-ToM's volatility parameter
}
\references{
(Daunizeau 2014)
}
\author{
K. Enevoldsen & P. Waade
}
