% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{basic_variance_update}
\alias{basic_variance_update}
\title{basic_variance_update}
\usage{
basic_variance_update(prev_hidden_states, params)
}
\arguments{
\item{prev_hidden_states}{a list structure containing the states from last round}

\item{params}{a list structure containing 0-ToM's volatility parameter}
}
\value{
An updated estimate variance
}
\description{
#0-ToM updates the variance of its parameter estimate
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
