% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{compete}
\alias{compete}
\title{compete}
\usage{
compete(player1, player2, p1_params, p2_params, p1_hidden_states = NA,
  p2_hidden_states = NA, p_matrix, n_rounds = 10)
}
\arguments{
\item{player1}{player 1's strategy function}

\item{player2}{player 2's strategy function}

\item{p1_params}{a list structure containing player 1's parameter values}

\item{p2_params}{a list structure containing player 2's parameter values}

\item{p1_hidden_states}{player 1's starting states, i.e. priors}

\item{p2_hidden_states}{player 2's starting states, i.e. priors}

\item{p_matrix}{a given 2-by-2 payoff matrix}

\item{n_rounds}{number of rounds to compete}
}
\value{
A dataframe containing the two agents' choices, rewards and states, for each round
}
\description{
Makes two agents play against each other a set number of rounds
}
\references{

}
\author{
K. Enevoldsen & P. Waade
}
