% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{create_agents}
\alias{create_agents}
\title{create_agents}
\usage{
create_agents(strategies, match_up = "RR", prepare_with_defaults = T)
}
\arguments{
\item{strategies}{A list of strategies given as a vector}

\item{match_up}{a string argument deciding which tournament structure is used}
}
\value{

}
\description{
Creates a dataframe with agents and their opponents.
If prepare_with_defaults is set to TRUE, this also prepares their parameter values and priors, which can fed to the compete function.
If prepare_with_defaults is set to FALSE, the output needs to be inserted into the prepare_df function.

#implemented agent strategies
RB: Random Bias
TFT: Tit for Tat
WSLS: Win-stay Loose-switch
SoftmaxTitTat: A tit for tat using a softmax function as well as an utility function
k-ToM: a k-level theory of mind agent. Note that you input the desired k-level in place of k, e.g. 1-ToM

implemented matchup types include
RR: Round robin, every strategy in the strategies input matched up against every other strategy
random: The strategies in the strategies input is randomly matched up
half_half: the first half is matched up with the second half
}
\references{

}
\author{
K. Enevoldsen & P. Waade
}
