% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{decision_function}
\alias{decision_function}
\title{decision_function}
\usage{
decision_function(hidden_states, params, player, level, p_matrix)
}
\arguments{
\item{hidden_states}{k-ToM's updated estimates}

\item{params}{a list structure containing k-ToM's volatility and behavioural temperature parameters}

\item{player}{k-ToM's player role, i.e. which side of the payoff matrix is used}

\item{level}{k-ToM's sophistication level k}

\item{p_matrix}{a given 2-by-2 payoff matrix}
}
\value{
k-ToM's own choice probability
}
\description{
k-ToM's decision function, where it calculates its own choice probability based on the updated estimates
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
