% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{expected_payoff_difference}
\alias{expected_payoff_difference}
\title{Expected payoff function}
\usage{
expected_payoff_difference(p_op_1, player, p_matrix)
}
\arguments{
\item{p_op_1}{estimated choice probability of the opponent}

\item{player}{the current player}

\item{p_matrix}{a given payoff matrix}
}
\value{
An expected payoff difference for 1
}
\description{
Returns the value difference between choosing 1 and 0 over opponent chocie scenarios, each scenario weighted by the probability of the opponent's choice
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
