% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{fetch_p_matrix}
\alias{fetch_p_matrix}
\title{Fetch payoff matrix}
\usage{
fetch_p_matrix(game, custom = c(0, 0, 0, 0, 0, 0, 0, 0))
}
\value{
A payoff matrix
}
\description{
Makes a payoff matrix for use in the simulation.
The following string commands are for getting pre-made matrices:
staghunt" - the stag hunt game
penny_cooperative" - a cooperative penny matching game
penny_competitive" - a competitive penny matching game
prisoners_dilemma" - the prisoner's dilemma game
party" - the party dilemma game
sexes" - the battle of the sexes game
chicken" - the chicken game
deadlock" - a deadlock game
custom - for specifying a custom game
}
\references{

}
\author{
K. Enevoldsen & P. Waade
}
