% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{k_ToM}
\alias{k_ToM}
\title{The k-ToM function}
\usage{
k_ToM(params = "default", hidden_states, player, level = NULL,
  p_matrix, choice_self = NULL, choice_op = NULL,
  return_hidden_states = T)
}
\arguments{
\item{params}{a list structure containing k-ToM's volatility parameter, the dummy variable which decides which parameter estimates are affected by volatility, and the behavioural temperature. If a string is inputted, default values are used.}

\item{hidden_states}{the estimates from last round}

\item{player}{k-ToM's player role, i.e. which side of the payoff matrix is used}

\item{level}{k-ToM's sophistication level k}

\item{p_matrix}{a given 2-by-2 payoff matrix}

\item{choice_self}{k-ToM's choice from last round}

\item{choice_op}{opponent's choice from last round}
}
\value{
A list structure containing k-ToM's choice and updated estimates
}
\description{
The full k-ToM function. First it first updates level probability, choice probability, parameter and gradient estimates. Then it calculates the estimated choice proability of the opponent, and calculates its own choice probability in response. The function also contains the simpler 0-ToM strategy, which only updates opponent's choice probability, and reacts.
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
