% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{p_op_1_k_approx_fun}
\alias{p_op_1_k_approx_fun}
\title{p_op_1_k_approx_fun}
\usage{
p_op_1_k_approx_fun(prev_hidden_states, level)
}
\arguments{
\item{prev_hidden_states}{a list structure containing the states from last round}

\item{level}{k-ToM's sophistication level}
}
\value{
An approximation of the estimated choice probability of last round.
}
\description{
Approximates the estimated choice probability of the opponent on the previous round. A semi-analytical approximation derived in Daunizeau, J. (2017)
}
\references{
Daunizeau, J. (2017)
}
\author{
K. Enevoldsen & P. Waade
}
