% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{p_op_1_k_fun}
\alias{p_op_1_k_fun}
\title{p_op_1_k_fun}
\usage{
p_op_1_k_fun(hidden_states, params)
}
\arguments{
\item{hidden_states}{k-ToM's updated estimates for this round}
}
\value{
The estimated choice probabilities of the opponent, for each possible opponent level
}
\description{
k-ToM combines the mean choice probability estimate and the variances of its parameter estimates into a final choice probability estimate.
NB: this is the function taken from the VBA package (Daunizeau 2014), which does not use k-ToM's volatility parameter
}
\references{
Daunizeau (2014)
}
\author{
K. Enevoldsen & P. Waade
}
