% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{parameter_mean_update}
\alias{parameter_mean_update}
\title{parameter_mean_update}
\usage{
parameter_mean_update(prev_hidden_states, choices, p_k, variance)
}
\arguments{
\item{prev_hidden_states}{A list strucutre containing the states from the last round}

\item{choices}{a vector of 1 own choice and 2 opponent's choice from last round}

\item{p_k}{a vector of level probabilities}

\item{variance}{k-ToM's variance on parameter estimates}
}
\value{
Updated means of k-ToM's parameter estimates
}
\description{
k-ToM updates the mean of its parameter estimates
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
