% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{parameter_variance_update}
\alias{parameter_variance_update}
\title{parameter_variance_update}
\usage{
parameter_variance_update(prev_hidden_states, params, p_k)
}
\arguments{
\item{prev_hidden_states}{A list strucutre containing the states from the last round}

\item{params}{A list structure containing k-ToM's volatility parameter and a dummy parameter which decides which parmaeter estimates are affected by volatility}

\item{p_k}{A vector of level probabilities}
}
\value{
Updated variances of k-ToM's parameter estimates
}
\description{
k-ToM updates the variance of its parameter estimates
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
