% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_agents.R
\name{PlaySelf}
\alias{PlaySelf}
\title{Play self}
\usage{
PlaySelf(params = NULL, hidden_states = NULL, player = NULL,
  p_matrix = NULL, choice_self = NULL, choice_op = NULL,
  return_hidden_states = F, messages = T)
}
\arguments{
\item{player}{the current player role}

\item{p_matrix}{the current payoff matric}

\item{choice_self}{player's own choice on last round}

\item{choice_op}{opponent's choice last round}
}
\value{
player's choice
}
\description{
Lets a human player make a choice
}
\references{

}
\author{
K. Enevoldsen & P. Waade
}
