% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{prepare}
\alias{prepare}
\title{prepare}
\usage{
prepare(strategy_string, RB = c(prop_mean = 0.8, prop_sd = 0.1),
  TFT = c(prob_mean = 0.9, prob_sd = 0.1), WSLS = c(stay_mean = 0.9,
  stay_sd = 0.1, switch_mean = 0.9, switch_sd = 0.1),
  k_ToM = c(volatility_mean = -2, volatility_sd = 0, temperature_mean =
  -10, temperature_sd = 0), k_ToM_priors = "default")
}
\arguments{
\item{strategy_string}{A string of the strategy used by the agent.}

\item{RB}{A vector of mean and standard deviation of the choice probability parameter (used for RB)}

\item{TFT}{A vector of mean and standard deviation of the choice probability parameter (used for TFT)}

\item{WSLS}{A vector of mean and standard deviation of the win stay probability and the loose shift parameters  (used for WSLS)}

\item{k_ToM}{A vector of mean and standard deviation of the volatility and behavioural temperature parameters (used for k-ToM)}

\item{k_ToM_priors}{A list structure containing priors for all of k-ToM's estimates. This is fed to the red_prepare_k_tom function. If a string is specified, defaults will be used}
}
\value{
A list structure containing the sampled parameter values of the agent and the states for the first round (if any).
}
\description{
Function for generating parameter values and priors for an agent. Parameter values are specified as vectors, where the first number is the mean and the second number is the standard deviation of the normal distribution they are sampled from. If no values are specified, defaults are used. Only the values for the strategy specified are returned.
Implemented agent strategies:
RB: Random Bias
TFT: Tit for Tat
WSLS: Win-stay Loose-switch
k-ToM: a k-level theory of mind agent. Note that you input the desired k-level in place of k, e.g. 1-ToM
}
\references{

}
\author{
K. Enevoldsen & P. Waade
}
