% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{prepare_df}
\alias{prepare_df}
\title{prepare_df}
\usage{
prepare_df(agent_df_unprep, args = NULL)
}
\arguments{
\item{agent_df_unprep}{a tibble with agent strategies matched up, but without set parameter values and prios. Created by the prepare_agents function}

\item{args}{a string vector of arguments to pass to the prepare function}
}
\value{
A dataframe with agents matched up and with parameter values and priors, to be fed to the compete function
}
\description{
Sets parameter values and priors for agents in a matched-up dataframe, to be fed to the compete function.
}
\references{

}
\author{
K. Enevoldsen & P. Waade
}
