% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_agents.R
\name{RB}
\alias{RB}
\title{Random Bias (RB)}
\usage{
RB(params, hidden_states = NULL, player = NULL, p_matrix = NULL,
  choice_self = NULL, choice_op = NULL, return_hidden_states = F)
}
\arguments{
\item{params}{a list of 1 element, RB's choice probability parameter}
}
\value{
RB's choice
}
\description{
RB: A Random Bias strategy choice. Selects randomly with a given probability
}
\references{

}
\author{
K. Enevoldsen & P. Waade
}
