% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{rec_learning_function}
\alias{rec_learning_function}
\title{rec_learning_function}
\usage{
rec_learning_function(prev_hidden_states, params, choices, level, player,
  p_matrix)
}
\arguments{
\item{prev_hidden_states}{a list structure containing the states from last round}

\item{params}{a list structure containing k-ToM's volatility parameter}

\item{choices}{a vector of 1 own choice and 2 opponent's choice from last round}

\item{level}{k-ToM's sophistication level}

\item{player}{k-ToM's player role, i.e. which side of the payoff matrix is used}

\item{p_matrix}{a given 2-by-2 payoff matrix}
}
\value{
A list structure containing the updates estimates by k-ToM and all the recursively simulated opponents
}
\description{
k-ToM's learning function, where it updates its level probability, parameter, choice probability and gradient etimates. This is called recursively.
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
