% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{rec_prepare_k_ToM}
\alias{rec_prepare_k_ToM}
\title{rec_prepare_k_ToM}
\usage{
rec_prepare_k_ToM(level, priors = "default")
}
\arguments{
\item{level}{k-ToM's sophistication level k}

\item{priors}{a list structure containing all the prior values for k-ToM. If a string is given, default values are used.}
}
\value{
A list structure which is ued as the previous estimates in k-ToM's first round.
}
\description{
A recursive function for preparing the list structure containing estimates which is used in the k-ToM function.
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
