% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{softmax}
\alias{softmax}
\title{Softmax function}
\usage{
softmax(e_payoff_diff, behavioural_temperature)
}
\arguments{
\item{e_payoff_diff}{An expected payoff difference for 1}

\item{behavioural_temperature}{A randomizing temperature parameter}
}
\value{
A choice probability for 1
}
\description{
Returns a choice probability for 1 given an expected payoff
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
