% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_agents.R
\name{TFT}
\alias{TFT}
\title{Tit For Tat (TFT)}
\usage{
TFT(params, hidden_states = NULL, player = NULL, p_matrix = NULL,
  choice_self, choice_op, return_hidden_states = F)
}
\arguments{
\item{params}{list of 1 element: TFT's choice probability parameter}
}
\value{
TFT's choice
}
\description{
A probabilistic Tit for Tat strategy. Copies the opponent's last choice with a given probability.
}
\references{
Shelling (1981)
}
\author{
K. Enevoldsen & P. Waade
}
