% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{U}
\alias{U}
\title{Utility function}
\usage{
U(c_self, c_op, player, p_matrix)
}
\arguments{
\item{c_self}{Player's own choice}

\item{c_op}{Opponent's choice}

\item{player}{Which side of the payof matrix is used. 0: first player, 1: second player}

\item{p_matrix}{A given 2-by-2 payoff matrix}
}
\value{
The reward for the specified player
}
\description{
Returns the reward for a player given a payoff matrix and choices
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
