% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToM.R
\name{update_pk}
\alias{update_pk}
\title{update_pk}
\usage{
update_pk(prev_hidden_states, choices, p_op_1_k_approx)
}
\arguments{
\item{prev_hidden_states}{A list structure containing the states of last trial}

\item{choices}{a vector of 1 own choice and 2 opponent's choice from last round}

\item{p_op_1_k_approx}{an approximated opponent choice probability form last round}
}
\value{
The updated level probabilities
}
\description{
Updates k-ToM's estimated probability of the opponent having each possible sophistication level.
}
\references{
Devaine et al. (2014a, 2014b, 2017)
}
\author{
K. Enevoldsen & P. Waade
}
