% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_agents.R
\name{WSLS}
\alias{WSLS}
\title{Win-stay loose-switch (WSLS)}
\usage{
WSLS(params, hidden_states = NULL, player, p_matrix, choice_self,
  choice_op, return_hidden_states = F)
}
\arguments{
\item{params}{A list of two elements, WSLS's win stay probability parameter and loose shift probability parameter}
}
\value{
WSLS's choice
}
\description{
A probabilistic Win-stay Loose-shift strategy. Copies itself if it won last round, does the opposite if it lost, both with given probabilities
}
\references{
Nowak, M., & Sigmund, K. (1993). A strategy of win-stay, lose-shift that outperforms tit-for-tat in the Prisoner's Dilemma game. Nature, 364(6432), 56.
}
\author{
K. Enevoldsen & P. Waade
}
